+++ service/multichain/opensea/opensea.go
@@ -196,131 +196,6 @@ func (p *Provider) ValidateTokensForWallet(context.Context, persist.Address, boo
 	return nil
 }
 
-// UpdateAssetsForAcc is a pipeline for getting assets for an account
-func UpdateAssetsForAcc(pCtx context.Context, pUserID persist.DBID, pOwnerWalletAddresses []persist.EthereumAddress,
-	nftRepo persist.NFTRepository, userRepo persist.UserRepository, collRepo persist.CollectionTokenRepository, galleryRepo persist.GalleryTokenRepository, backupRepo persist.BackupRepository) error {
-
-	err := galleryRepo.RefreshCache(pCtx, pUserID)
-	if err != nil {
-		return err
-	}
-
-	galleries, err := galleryRepo.GetByUserID(pCtx, pUserID)
-	if err != nil {
-		return err
-	}
-
-	for _, gallery := range galleries {
-		err = backupRepo.Insert(pCtx, gallery)
-		if err != nil {
-			return err
-		}
-	}
-
-	user, err := userRepo.GetByID(pCtx, pUserID)
-	if err != nil {
-		return fmt.Errorf("failed to get user by id %s: %w", pUserID, err)
-	}
-	if len(pOwnerWalletAddresses) == 0 {
-		pOwnerWalletAddresses = persist.WalletsToEthereumAddresses(user.Wallets)
-	}
-
-	ids, err := UpdateAssetsForWallet(pCtx, pOwnerWalletAddresses, nftRepo)
-	if err != nil {
-		if e, ok := err.(ErrNoAssetsForWallets); ok {
-			logrus.Debugf("no assets found for wallets %v", e.Wallets)
-		} else {
-			return err
-		}
-	}
-
-	// ensure NFTs that a user used to own are no longer in their gallery
-	if err := collRepo.ClaimNFTs(pCtx, pUserID, pOwnerWalletAddresses, persist.CollectionTokenUpdateTokensInput{Tokens: ids}); err != nil {
-		return fmt.Errorf("failed to claim NFTs: %w", err)
-	}
-	if err := collRepo.RemoveNFTsOfOldAddresses(pCtx, pUserID); err != nil {
-		return fmt.Errorf("failed to remove NFTs of old addresses: %w", err)
-	}
-
-	return nil
-}
-
-// UpdateAssetsForWallet is a pipeline for getting assets for a wallet
-func UpdateAssetsForWallet(pCtx context.Context, pOwnerWalletAddresses []persist.EthereumAddress, nftRepo persist.NFTRepository) ([]persist.DBID, error) {
-	var returnErr error
-	asDBNfts, err := fetchAssetsForWallets(pCtx, pOwnerWalletAddresses, nftRepo)
-	if err != nil {
-		if e, ok := err.(ErrNoAssetsForWallets); ok {
-			returnErr = e
-		} else {
-			return nil, fmt.Errorf("failed to fetch assets for wallets %v: %w", pOwnerWalletAddresses, err)
-		}
-	}
-	logrus.Debugf("found %d assets for wallets %v", len(asDBNfts), pOwnerWalletAddresses)
-
-	ids, err := nftRepo.BulkUpsert(pCtx, asDBNfts)
-	if err != nil {
-		return nil, fmt.Errorf("failed to bulk upsert NFTs: %w", err)
-	}
-	logrus.Debugf("bulk upserted %d NFTs", len(ids))
-	return ids, returnErr
-}
-
-func fetchAssetsForWallets(pCtx context.Context, pWalletAddresses []persist.EthereumAddress, nftRepo persist.NFTRepository) ([]persist.NFT, error) {
-	result := []persist.NFT{}
-	nftsChan := make(chan []persist.NFT)
-	errChan := make(chan error)
-	for _, walletAddress := range pWalletAddresses {
-		go func(wa persist.EthereumAddress) {
-			assets, err := FetchAssetsForWallet(pCtx, wa, "", 0, nil)
-			if err != nil {
-				errChan <- fmt.Errorf("failed to fetch assets for wallet %s: %s", wa, err)
-				return
-			}
-			if len(assets) == 0 {
-				time.Sleep(time.Second)
-				assets, err = FetchAssetsForWallet(pCtx, wa, "", 0, nil)
-				if err != nil {
-					errChan <- fmt.Errorf("failed to fetch assets for wallet %s: %s", wa, err)
-					return
-				}
-				if len(assets) == 0 {
-					errChan <- ErrNoAssetsForWallets{Wallets: []persist.EthereumAddress{wa}}
-					return
-				}
-			}
-			asGlry, err := assetsToNFTs(pCtx, wa, assets, nftRepo)
-			if err != nil {
-				errChan <- fmt.Errorf("failed to convert opensea assets to db nfts: %s", err)
-				return
-			}
-			nftsChan <- asGlry
-		}(walletAddress)
-	}
-
-	var notFoundError error
-	for i := 0; i < len(pWalletAddresses); i++ {
-		select {
-		case nfts := <-nftsChan:
-			result = append(result, nfts...)
-		case err := <-errChan:
-			if e, ok := err.(ErrNoAssetsForWallets); ok {
-				if notFoundError == nil {
-					notFoundError = e
-				} else {
-					it := notFoundError.(ErrNoAssetsForWallets)
-					it.Wallets = append(it.Wallets, e.Wallets...)
-					notFoundError = it
-				}
-			} else {
-				return nil, err
-			}
-		}
-	}
-
-	return result, notFoundError
-}
-
 // FetchAssetsForWallet recursively fetches all assets for a wallet
 func FetchAssetsForWallet(pCtx context.Context, pWalletAddress persist.EthereumAddress, pCursor string, retry int, alreadyReceived map[int]string) ([]Asset, error) {
 
@@ -593,54 +468,6 @@ func contractToContract(ctx context.Context, openseaContract Contract, ethClient
 		CreatorAddress: persist.Address(openseaContract.Collection.PayoutAddress),
 		LatestBlock:    persist.BlockNumber(block),
 	}, nil
-}
-
-func assetsToNFTs(pCtx context.Context, pWalletAddress persist.EthereumAddress, openseaNfts []Asset, nftRepo persist.NFTRepository) ([]persist.NFT, error) {
-
-	nfts := make([]persist.NFT, len(openseaNfts))
-	nftChan := make(chan persist.NFT)
-	for _, openseaNft := range openseaNfts {
-		go func(nft Asset) {
-			nftChan <- openseaToDBNft(pCtx, pWalletAddress, nft, nftRepo)
-		}(openseaNft)
-	}
-	for i := 0; i < len(openseaNfts); i++ {
-		nfts[i] = <-nftChan
-	}
-	return nfts, nil
-}
-
-func openseaToDBNft(pCtx context.Context, pWalletAddress persist.EthereumAddress, nft Asset, nftRepo persist.NFTRepository) persist.NFT {
-
-	result := persist.NFT{
-		OwnerAddress:         pWalletAddress,
-		MultipleOwners:       nft.Owner.Address == persist.ZeroAddress,
-		Name:                 persist.NullString(nft.Name),
-		Description:          persist.NullString(nft.Description),
-		ExternalURL:          persist.NullString(nft.ExternalURL),
-		ImageURL:             persist.NullString(nft.ImageURL),
-		CreatorAddress:       nft.Creator.Address,
-		AnimationURL:         persist.NullString(nft.AnimationURL),
-		OpenseaTokenID:       persist.TokenID(nft.TokenID.ToBase16()),
-		OpenseaID:            persist.NullInt64(nft.ID),
-		TokenCollectionName:  persist.NullString(nft.Collection.Name),
-		ImageThumbnailURL:    persist.NullString(nft.ImageThumbnailURL),
-		ImagePreviewURL:      persist.NullString(nft.ImagePreviewURL),
-		ImageOriginalURL:     persist.NullString(nft.ImageOriginalURL),
-		TokenMetadataURL:     persist.NullString(nft.TokenMetadataURL),
-		Contract:             nft.Contract,
-		AcquisitionDateStr:   persist.NullString(nft.AcquisitionDateStr),
-		CreatorName:          persist.NullString(nft.Creator.User.Username),
-		AnimationOriginalURL: persist.NullString(nft.AnimationOriginalURL),
-	}
-
-	dbNFT, _ := nftRepo.GetByOpenseaID(pCtx, persist.NullInt64(nft.ID), pWalletAddress)
-	if dbNFT.ID != "" {
-		result.ID = dbNFT.ID
-	}
-
-	return result
-
 }
 
 // VerifySignature will verify a signature using all available methods (eth_sign and personal_sign)
