+++ service/persist/gallery.go
@@ -5,11 +5,11 @@ import (
 	"fmt"
 )
 
-// GalleryTokenDB represents a group of collections of NFTs in the database.
+// GalleryDB represents a group of collections of NFTs in the database.
 // Collections of NFTs will be represented as a list of collection IDs creating
 // a join relationship in the database
 // This struct will only be used in database operations
-type GalleryTokenDB struct {
+type GalleryDB struct {
 	Version      NullInt32       `json:"version"` // schema version for this model
 	ID           DBID            `json:"id" binding:"required"`
 	CreationTime CreationTime    `json:"created_at"`
@@ -20,19 +20,19 @@ type GalleryTokenDB struct {
 	Collections []DBID `json:"collections"`
 }
 
-// GalleryToken represents a group of collections of NFTS in the application.
+// Gallery represents a group of collections of NFTS in the application.
 // Collections are represented as structs instead of IDs
 // This struct will be decoded from a find database operation and used throughout
 // the application where GalleryDB is not used
-type GalleryToken struct {
+type Gallery struct {
 	Version      NullInt32       `json:"version"` // schema version for this model
 	ID           DBID            `json:"id" binding:"required"`
 	CreationTime CreationTime    `json:"created_at"`
 	Deleted      NullBool        `json:"-"`
 	LastUpdated  LastUpdatedTime `json:"last_updated"`
 
-	OwnerUserID DBID              `json:"owner_user_id"`
-	Collections []CollectionToken `json:"collections"`
+	OwnerUserID DBID         `json:"owner_user_id"`
+	Collections []Collection `json:"collections"`
 }
 
 // GalleryTokenUpdateInput represents a struct that is used to update a gallery's list of collections in the databse
@@ -42,14 +42,14 @@ type GalleryTokenUpdateInput struct {
 	Collections []DBID `json:"collections"`
 }
 
-// GalleryTokenRepository is an interface for interacting with the gallery persistence layer
-type GalleryTokenRepository interface {
-	Create(context.Context, GalleryTokenDB) (DBID, error)
+// GalleryRepository is an interface for interacting with the gallery persistence layer
+type GalleryRepository interface {
+	Create(context.Context, GalleryDB) (DBID, error)
 	Update(context.Context, DBID, DBID, GalleryTokenUpdateInput) error
 	UpdateUnsafe(context.Context, DBID, GalleryTokenUpdateInput) error
 	AddCollections(context.Context, DBID, DBID, []DBID) error
-	GetByUserID(context.Context, DBID) ([]GalleryToken, error)
-	GetByID(context.Context, DBID) (GalleryToken, error)
+	GetByUserID(context.Context, DBID) ([]Gallery, error)
+	GetByID(context.Context, DBID) (Gallery, error)
 	RefreshCache(context.Context, DBID) error
 }
